+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix a crash when inputting a too large amount as the stake for delegation or baking.
+- Fix validation of token metadata when looking for new tokens.
 - Fix an issue where another tokens image was sometimes shown for tokens without an image.
 - Fix when searching for a token, the "no tokens found" display did not appear.
 - Fixed number of issues with 
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   5. NFT token details screen should not have Decimal section.
   6. Token image size is very small, contents almost invisible.
 - Fixed issue where owned tokens are not reflected in search result list and details screens
+- When adding a new token the token details would in some cases show information from another token.
 
 ### Changed